refactor: common의 export 방식 변경하기
