+++ app/my-profile/activity.tsx
@@ -4,7 +4,7 @@ import { View } from 'react-native';
 import { ScrollView } from "react-native-gesture-handler";
 
 import { ImageButton } from '@/components/my-profile/activity';
-import MenuGroupList from '@/components/my-profile/common/menu-group-list';
+import { MenuGroupList } from '@/components/my-profile/common';
 
 
 // TODO: 가상 데이터를 실제 데이터로 교체하기
