+++ components/my-profile/common/index.ts
@@ -1,0 +1,4 @@
+export { default as Button } from "./button";
+export { default as MenuGroupList } from "./menu-group-list";
+export { default as ProfileImageBox } from './profile-image-box';
+export { default as  UserActivitySummary } from './user-activity-summary';
